+++ packages/apollo-language-server/src/errors/validation.ts
@@ -1,7 +1,6 @@
 import {
   specifiedRules,
   NoUnusedFragmentsRule,
-  KnownDirectivesRule,
   GraphQLError,
   FieldNode,
   ValidationContext,
@@ -12,17 +11,30 @@ import {
   FragmentDefinitionNode,
   visit,
   visitWithTypeInfo,
-  visitInParallel
+  visitInParallel,
+  getLocation
 } from "graphql";
+
+import { TextEdit } from "vscode-languageserver";
 
 import { ToolError, logError } from "./logger";
 import { ValidationRule } from "graphql/validation/ValidationContext";
+import {
+  positionFromSourceLocation,
+  isFieldResolvedLocally
+} from "../utilities/source";
 
-const specifiedRulesToBeRemoved = [NoUnusedFragmentsRule, KnownDirectivesRule];
+export interface CodeActionInfo {
+  message: string;
+  edits: TextEdit[];
+}
+
+const specifiedRulesToBeRemoved = [NoUnusedFragmentsRule];
 
 export const defaultValidationRules: ValidationRule[] = [
   NoAnonymousQueries,
   NoTypenameAlias,
+  NoMissingClientDirectives,
   ...specifiedRules.filter(rule => !specifiedRulesToBeRemoved.includes(rule))
 ];
 
@@ -90,6 +102,69 @@ export function NoTypenameAlias(context: ValidationContext) {
           )
         );
       }
+    }
+  };
+}
+
+export function NoMissingClientDirectives(context: ValidationContext) {
+  const root = context.getDocument();
+  return {
+    Field(node: FieldNode) {
+      const parentType = context.getParentType();
+      const fieldDef = context.getFieldDef();
+
+      if (!parentType || !fieldDef) return;
+
+      const isClientType =
+        parentType.clientSchema &&
+        parentType.clientSchema.localFields &&
+        parentType.clientSchema.localFields.includes(fieldDef.name);
+
+      const isResolvedLocally = isFieldResolvedLocally(node, root);
+
+      if (isClientType && !isResolvedLocally) {
+        let extensions: { [key: string]: any } | null = null;
+        const nameLoc = node.name.loc;
+        if (nameLoc) {
+          let { source, end: locToInsertDirective } = nameLoc;
+          if (node.arguments && node.arguments.length !== 0) {
+            // must insert directive after field arguments
+            const endOfArgs = source.body.indexOf(")", locToInsertDirective);
+            locToInsertDirective = endOfArgs + 1;
+          }
+          const codeAction: CodeActionInfo = {
+            message: `Add @client directive to "${node.name.value}"`,
+            edits: [
+              TextEdit.insert(
+                positionFromSourceLocation(
+                  source,
+                  getLocation(source, locToInsertDirective)
+                ),
+                " @client"
+              )
+            ]
+          };
+          extensions = { codeAction };
+        }
+
+        context.reportError(
+          new GraphQLError(
+            `Local field "${node.name.value}" must have a @client directive`,
+            [node],
+            null,
+            null,
+            null,
+            null,
+            extensions
+          )
+        );
+      }
+
+      if (isClientType) {
+        return false;
+      }
+
+      return;
     }
   };
 }