+++ packages/apollo-language-server/src/server.ts
@@ -82,7 +82,7 @@ connection.onInitialize(async ({ capabilities, workspaceFolders }) => {
       hoverProvider: true,
       completionProvider: {
         resolveProvider: false,
-        triggerCharacters: ["..."]
+        triggerCharacters: ["...", "@"]
       },
       definitionProvider: true,
       referencesProvider: true,
@@ -91,6 +91,7 @@ connection.onInitialize(async ({ capabilities, workspaceFolders }) => {
       codeLensProvider: {
         resolveProvider: false
       },
+      codeActionProvider: true,
       executeCommandProvider: {
         commands: []
       },
@@ -186,18 +187,30 @@ connection.onWorkspaceSymbol((params, token) =>
 );
 
 connection.onCompletion(
-  debounceHandler((params, token) =>
-    languageProvider.provideCompletionItems(
-      params.textDocument.uri,
-      params.position,
-      token
-    )
+  debounceHandler(
+    (params, token) =>
+      languageProvider.provideCompletionItems(
+        params.textDocument.uri,
+        params.position,
+        token
+      ),
+    false
   )
 );
 
 connection.onCodeLens(
   debounceHandler((params, token) =>
     languageProvider.provideCodeLenses(params.textDocument.uri, token)
+  )
+);
+
+connection.onCodeAction(
+  debounceHandler((params, token) =>
+    languageProvider.provideCodeAction(
+      params.textDocument.uri,
+      params.range,
+      token
+    )
   )
 );